requirements.txt
django
aiogram
djangorestframework
requests